(BSR)[PRO] refactor: cleanup cells import paths
