--- pro/src/pages/Offers/Offers/OfferItem/Cells/CheckboxCell/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CheckboxCell'
