+++ pro/src/pages/Offers/Offers/OfferItem/Cells/DuplicateOfferCell/DuplicateOfferCell.tsx
@@ -19,7 +19,7 @@ import DuplicateOfferDialog from './DuplicateOfferDialog'
 
 export const LOCAL_STORAGE_HAS_SEEN_MODAL_KEY = 'DUPLICATE_OFFER_MODAL_SEEN'
 
-const DuplicateOfferCell = ({
+export const DuplicateOfferCell = ({
   offerId,
   isShowcase,
 }: {
@@ -96,4 +96,3 @@ const DuplicateOfferCell = ({
     </>
   )
 }
-export default DuplicateOfferCell
