+++ pro/src/pages/Offers/Offers/OfferItem/Cells/EditOfferCell.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import penIcon from 'icons/full-edit.svg'
 import ListIconButton from 'ui-kit/ListIconButton/ListIconButton'
 
-import styles from '../../OfferItem.module.scss'
+import styles from '../OfferItem.module.scss'
 
 interface EditOfferCellProps {
   editionOfferLink: string
