+++ pro/src/pages/Offers/Offers/OfferItem/Cells/EditStocksCell.tsx
