--- pro/src/pages/Offers/Offers/OfferItem/Cells/OfferInstitutionCell/index.ts
@@ -1,1 +1,0 @@
-export { default } from './OfferInstitutionCell'
