--- pro/src/pages/Offers/Offers/OfferItem/Cells/OfferNameCell/index.ts
@@ -1,1 +1,0 @@
-export { default } from './OfferNameCell'
