+++ pro/src/pages/Offers/Offers/OfferItem/Cells/OfferNameCell/utils.ts
