+++ pro/src/pages/Offers/Offers/OfferItem/Cells/OfferStatusCell.tsx
@@ -3,12 +3,14 @@ import React from 'react'
 import { OfferStatus } from 'apiClient/v1'
 import { StatusLabel } from 'components/StatusLabel/StatusLabel'
 
-import styles from '../../OfferItem.module.scss'
+import styles from '../OfferItem.module.scss'
 
-const OfferStatusCell = ({ status }: { status: OfferStatus }) => (
+interface OfferStatusCellProps {
+  status: OfferStatus
+}
+
+export const OfferStatusCell = ({ status }: OfferStatusCellProps) => (
   <td className={styles['status-column']}>
     <StatusLabel status={status} />
   </td>
 )
-
-export default OfferStatusCell
