--- pro/src/pages/Offers/Offers/OfferItem/Cells/OfferStatusCell/index.ts
@@ -1,1 +1,0 @@
-export { default } from './OfferStatusCell'
