+++ pro/src/pages/Offers/Offers/OfferItem/Cells/OfferVenueCell.tsx
@@ -3,14 +3,16 @@ import React from 'react'
 import { ListOffersVenueResponseModel } from 'apiClient/v1'
 import { computeVenueDisplayName } from 'repository/venuesService'
 
-import styles from '../../OfferItem.module.scss'
+import styles from '../OfferItem.module.scss'
 
-const OfferVenueCell = ({ venue }: { venue: ListOffersVenueResponseModel }) => {
+interface OfferVenueCellProps {
+  venue: ListOffersVenueResponseModel
+}
+
+export const OfferVenueCell = ({ venue }: OfferVenueCellProps) => {
   return (
     <td className={styles['venue-column']}>
       {venue && computeVenueDisplayName(venue)}
     </td>
   )
 }
-
-export default OfferVenueCell
