--- pro/src/pages/Offers/Offers/OfferItem/Cells/OfferVenueCell/index.ts
@@ -1,1 +1,0 @@
-export { default } from './OfferVenueCell'
